+++ src/main/java/com/my_project/niit_final_project/entities/ProductType.java
@@ -20,7 +20,7 @@ public class ProductType {
     String name;
 
     @Column(name = "status")
-    Long status;
+    Integer status;
 
     @OneToMany(mappedBy="productType",fetch = FetchType.LAZY)
     Collection<Category> categories;
